12: use long

+++ app/src/main/java/cs2/Program12.java
@@ -7,13 +7,13 @@ import java.util.Scanner;
 // Program 12
 
 // This program will have a recursive method called
-// power (int base, int exponent) that will eventually
+// power (long base, long exponent) that will eventually
 // return 3*3*3*3 if the base is 3 and the exponent is 4
 
 // Assume exponent is always >1
 
 public class Program12 {
-	public static int power(int base, int exponent) {
+	public static long power(long base, long exponent) {
 		if (exponent == 1) {
 			return base;
 		}
@@ -25,14 +25,14 @@ public class Program12 {
 
 		while (true) {
 			System.out.print("Enter a base (-1 to quit): ");
-			int base = sc.nextInt();
+			long base = sc.nextInt();
 
 			if (base == -1) {
 				break;
 			}
 
 			System.out.print("Enter an exponent (-1 to quit): ");
-			int exponent = sc.nextInt();
+			long exponent = sc.nextInt();
 
 			if (exponent == -1) {
 				break;
